Refactor duplicate modifier quick fixes into common reusable classes
