+++ .github/workflows/lsp4jakarta.yaml
@@ -4,7 +4,7 @@ on:
   push:
     branches: [main]
   pull_request:
-    branches: [main]
+    branches: [main, lsp4jakarta-0.2.7-integration-branch]
 
 jobs:
   build:
