+++ jakarta.jdt/org.eclipse.lsp4jakarta.jdt.core/src/main/java/org/eclipse/lsp4jakarta/jdt/core/java/codeaction/RemoveAbstractModifierQuickFix.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
-* Copyright (c) 2023 IBM Corporation and others.
+* Copyright (c) 2026 IBM Corporation and others.
 *
 * This program and the accompanying materials are made available under the
 * terms of the Eclipse Public License v. 2.0 which is available at
@@ -10,22 +10,21 @@
 * Contributors:
 *     IBM Corporation - initial implementation
 *******************************************************************************/
-package org.eclipse.lsp4jakarta.jdt.internal.persistence;
+package org.eclipse.lsp4jakarta.jdt.core.java.codeaction;
 
 import org.eclipse.lsp4jakarta.commons.codeaction.ICodeActionId;
 import org.eclipse.lsp4jakarta.commons.codeaction.JakartaCodeActionId;
-import org.eclipse.lsp4jakarta.jdt.core.java.codeaction.RemoveModifierConflictQuickFix;
 
 /**
- * Removes the final modifier from the declaring element.
+ * Removes the abstract modifier from the declaring element.
  */
-public class RemoveFinalModifierQuickFix extends RemoveModifierConflictQuickFix {
+public class RemoveAbstractModifierQuickFix extends RemoveModifierConflictQuickFix {
 
     /**
      * Constructor.
      */
-    public RemoveFinalModifierQuickFix() {
-        super("final");
+    public RemoveAbstractModifierQuickFix() {
+        super(false, "abstract");
     }
 
     /**
@@ -33,7 +32,7 @@ public class RemoveFinalModifierQuickFix extends RemoveModifierConflictQuickFix 
      */
     @Override
     public String getParticipantId() {
-        return RemoveFinalModifierQuickFix.class.getName();
+        return RemoveAbstractModifierQuickFix.class.getName();
     }
 
     /**
@@ -41,6 +40,6 @@ public class RemoveFinalModifierQuickFix extends RemoveModifierConflictQuickFix 
      */
     @Override
     protected ICodeActionId getCodeActionId() {
-        return JakartaCodeActionId.PersistenceRemoveFinalModifier;
+        return JakartaCodeActionId.RemoveAbstractModifier;
     }
 }
