+++ jakarta.jdt/org.eclipse.lsp4jakarta.jdt.core/src/main/java/org/eclipse/lsp4jakarta/jdt/core/java/codeaction/RemoveFinalModifierQuickFix.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
-* Copyright (c) 2023 IBM Corporation and others.
+* Copyright (c) 2026 IBM Corporation and others.
 *
 * This program and the accompanying materials are made available under the
 * terms of the Eclipse Public License v. 2.0 which is available at
@@ -10,21 +10,21 @@
 * Contributors:
 *     IBM Corporation - initial implementation
 *******************************************************************************/
-package org.eclipse.lsp4jakarta.jdt.internal.di;
+package org.eclipse.lsp4jakarta.jdt.core.java.codeaction;
 
+import org.eclipse.lsp4jakarta.commons.codeaction.ICodeActionId;
 import org.eclipse.lsp4jakarta.commons.codeaction.JakartaCodeActionId;
-import org.eclipse.lsp4jakarta.jdt.core.java.codeaction.RemoveModifierConflictQuickFix;
 
 /**
- * Removes the abstract modifier from the declaring element.
+ * Removes the final modifier from the declaring element.
  */
-public class RemoveAbstractModifierQuickFix extends RemoveModifierConflictQuickFix {
+public class RemoveFinalModifierQuickFix extends RemoveModifierConflictQuickFix {
 
     /**
      * Constructor.
      */
-    public RemoveAbstractModifierQuickFix() {
-        super(false, "abstract");
+    public RemoveFinalModifierQuickFix() {
+        super(false, "final");
     }
 
     /**
@@ -32,14 +32,14 @@ public class RemoveAbstractModifierQuickFix extends RemoveModifierConflictQuickF
      */
     @Override
     public String getParticipantId() {
-        return RemoveAbstractModifierQuickFix.class.getName();
+        return RemoveFinalModifierQuickFix.class.getName();
     }
 
     /**
      * {@inheritDoc}
      */
     @Override
-    protected JakartaCodeActionId getCodeActionId() {
-        return JakartaCodeActionId.DIRemoveAbstractModifier;
+    protected ICodeActionId getCodeActionId() {
+        return JakartaCodeActionId.RemoveFinalModifier;
     }
 }
