--- jakarta.jdt/org.eclipse.lsp4jakarta.jdt.core/src/main/java/org/eclipse/lsp4jakarta/jdt/internal/di/RemoveFinalModifierQuickFix.java
@@ -1,45 +1,0 @@
-/*******************************************************************************
-* Copyright (c) 2023 IBM Corporation and others.
-*
-* This program and the accompanying materials are made available under the
-* terms of the Eclipse Public License v. 2.0 which is available at
-* http://www.eclipse.org/legal/epl-2.0.
-*
-* SPDX-License-Identifier: EPL-2.0
-*
-* Contributors:
-*     IBM Corporation - initial implementation
-*******************************************************************************/
-package org.eclipse.lsp4jakarta.jdt.internal.di;
-
-import org.eclipse.lsp4jakarta.commons.codeaction.JakartaCodeActionId;
-import org.eclipse.lsp4jakarta.jdt.core.java.codeaction.RemoveModifierConflictQuickFix;
-
-/**
- * Removes the final modifier from the declaring element.
- */
-public class RemoveFinalModifierQuickFix extends RemoveModifierConflictQuickFix {
-
-    /**
-     * Constructor.
-     */
-    public RemoveFinalModifierQuickFix() {
-        super(false, "final");
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public String getParticipantId() {
-        return RemoveFinalModifierQuickFix.class.getName();
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    protected JakartaCodeActionId getCodeActionId() {
-        return JakartaCodeActionId.DIRemoveFinalModifier;
-    }
-}
