+++ jakarta.jdt/org.eclipse.lsp4jakarta.jdt.test/src/main/java/org/eclipse/lsp4jakarta/jdt/test/interceptor/InterceptorTest.java
@@ -12,14 +12,21 @@
 *******************************************************************************/
 package org.eclipse.lsp4jakarta.jdt.test.interceptor;
 
+import static org.eclipse.lsp4jakarta.jdt.test.core.JakartaForJavaAssert.assertJavaCodeAction;
 import static org.eclipse.lsp4jakarta.jdt.test.core.JakartaForJavaAssert.assertJavaDiagnostics;
+import static org.eclipse.lsp4jakarta.jdt.test.core.JakartaForJavaAssert.ca;
+import static org.eclipse.lsp4jakarta.jdt.test.core.JakartaForJavaAssert.createCodeActionParams;
 import static org.eclipse.lsp4jakarta.jdt.test.core.JakartaForJavaAssert.d;
+import static org.eclipse.lsp4jakarta.jdt.test.core.JakartaForJavaAssert.te;
 import java.util.Arrays;
 import org.eclipse.core.resources.IFile;
 import org.eclipse.core.runtime.Path;
 import org.eclipse.jdt.core.IJavaProject;
+import org.eclipse.lsp4j.CodeAction;
 import org.eclipse.lsp4j.Diagnostic;
 import org.eclipse.lsp4j.DiagnosticSeverity;
+import org.eclipse.lsp4j.TextEdit;
+import org.eclipse.lsp4jakarta.commons.JakartaJavaCodeActionParams;
 import org.eclipse.lsp4jakarta.commons.JakartaJavaDiagnosticsParams;
 import org.eclipse.lsp4jakarta.jdt.core.utils.IJDTUtils;
 import org.eclipse.lsp4jakarta.jdt.internal.core.ls.JDTUtilsLSImpl;
@@ -68,6 +75,11 @@ public class InterceptorTest extends BaseJakartaTest {
                                                DiagnosticSeverity.Error, "jakarta-interceptor", "InvalidInterceptorAnnotationOnAbstractClass");
 
         assertJavaDiagnostics(diagnosticsParams, IJDT_UTILS, invalidAbstractModifier);
+
+        JakartaJavaCodeActionParams codeActionParams = createCodeActionParams(uri, invalidAbstractModifier);
+        TextEdit te = te(5, 6, 5, 15, "");
+        CodeAction ca = ca(uri, "Remove the 'abstract' modifier", invalidAbstractModifier, te);
+        assertJavaCodeAction(codeActionParams, IJDT_UTILS, ca);
     }
 
     @Test
